# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/SafeMPC/mpc-service
  version: 0.1.0
paths:
...
  /v1/wallets:
    get:
      security:
      - Bearer: []
      description: 获取用户的所有钱包列表
      tags:
      - Wallets
      summary: 列出钱包
      operationId: getWallets
      parameters:
      - type: string
        description: 过滤区块链类型
        name: chain_type
        in: query
      - type: integer
        default: 20
        description: 每页数量
        name: limit
        in: query
      - type: integer
        default: 0
        description: 偏移量
        name: offset
        in: query
      responses:
        "200":
          description: 钱包列表
          schema:
            $ref: '#/definitions/listWalletsResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
    post:
      security:
      - Bearer: []
      description: 通过 DKG 创建 2-of-2 MPC 钱包，需要 WebAuthn 二次验证
      tags:
      - Wallets
      summary: 创建 MPC 钱包
      operationId: postCreateWallet
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/postCreateWalletPayload'
      responses:
        "201":
          description: 钱包创建会话已创建
          schema:
            $ref: '#/definitions/createWalletResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/publicHttpError'
        "401":
          description: 未授权或 WebAuthn 验证失败
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
  /v1/wallets/{walletId}:
    get:
      security:
      - Bearer: []
      description: 获取指定钱包的详细信息
      tags:
      - Wallets
      summary: 获取钱包详情
      operationId: getWallet
      parameters:
      - type: string
        description: 钱包 ID
        name: walletId
        in: path
        required: true
      responses:
        "200":
          description: 钱包详情
          schema:
            $ref: '#/definitions/walletResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: 钱包不存在
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /v1/wallets/{walletId}/balance:
    get:
      security:
      - Bearer: []
      description: 查询指定钱包在区块链上的余额
      tags:
      - Wallets
      summary: 查询钱包余额
      operationId: getWalletBalance
      parameters:
      - type: string
        name: walletId
        in: path
        required: true
      - type: string
        description: 区块链类型
        name: chain_type
        in: query
        required: true
      responses:
        "200":
          description: 余额信息
          schema:
            $ref: '#/definitions/walletBalanceResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: 钱包不存在
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
  /v1/wallets/{walletId}/sign:
    post:
      security:
      - Bearer: []
      description: 使用 MPC 钱包签名交易，需要 WebAuthn 二次验证
      tags:
      - Wallets
      summary: 签名交易
      operationId: postSignTransaction
      parameters:
      - type: string
        name: walletId
        in: path
        required: true
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/postSignTransactionPayload'
      responses:
        "200":
          description: 签名会话已创建
          schema:
            $ref: '#/definitions/signTransactionResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/publicHttpError'
        "401":
          description: 未授权或 WebAuthn 验证失败
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: 钱包不存在
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
  /v1/wallets/{walletId}/transactions:
    get:
      security:
      - Bearer: []
      description: 获取钱包的交易历史记录
      tags:
      - Wallets
      summary: 查询交易历史
      operationId: getWalletTransactions
      parameters:
      - type: string
        name: walletId
        in: path
        required: true
      - type: string
        description: 区块链类型
        name: chain_type
        in: query
        required: true
      - type: integer
        default: 20
        description: 每页数量
        name: limit
        in: query
      - type: integer
        default: 0
        description: 偏移量
        name: offset
        in: query
      responses:
        "200":
          description: 交易历史
          schema:
            $ref: '#/definitions/transactionsResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: 钱包不存在
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/publicHttpError'
definitions:
  createWalletResponse:
    type: object
    required:
    - wallet_id
    - dkg_session_id
    - status
    properties:
      dkg_session_id:
        description: DKG 会话 ID
        type: string
        format: uuid
      session_token:
        description: 用于连接 Signer 的鉴权 Token
        type: string
      signer_endpoints:
        description: Signer 节点 gRPC 地址列表
        type: array
        items:
          type: string
        example:
        - 192.168.1.100:9091
      status:
        description: 会话状态
        type: string
        enum:
        - pending
        - running
        - completed
        - failed
        example: pending
      wallet_id:
        description: 钱包 ID
        type: string
        format: uuid
...
  listWalletsResponse:
    type: object
    required:
    - wallets
    - total
    - limit
    - offset
    properties:
      limit:
        description: 每页数量
        type: integer
      offset:
        description: 偏移量
        type: integer
      total:
        description: 总数
        type: integer
      wallets:
        type: array
        items:
          $ref: '#/definitions/walletSummary'
...
  postCreateWalletPayload:
    type: object
    required:
    - algorithm
    - curve
    - chain_type
    properties:
      algorithm:
        description: 签名算法
        type: string
        enum:
        - ECDSA
        - EdDSA
        example: ECDSA
      chain_type:
        description: 区块链类型
        type: string
        enum:
        - ethereum
        - bitcoin
        - solana
        - polkadot
        example: ethereum
      curve:
        description: 椭圆曲线
        type: string
        enum:
        - secp256k1
        - ed25519
        example: secp256k1
      mobile_node_id:
        description: 移动端节点 ID (2-of-2 模式)
        type: string
        example: mobile-p1
      webauthn_assertion:
        $ref: '#/definitions/webAuthnAssertion'
...
  postSignTransactionPayload:
    type: object
    required:
    - message_hex
    - chain_type
    properties:
      chain_type:
        type: string
        enum:
        - ethereum
        - bitcoin
        - solana
        example: ethereum
      derivation_path:
        description: 派生路径（可选）
        type: string
        example: m/44'/60'/0'/0/0
      message_hex:
        description: 待签名的消息（hex）
        type: string
        example: 0xf86c...
      webauthn_assertion:
        $ref: '#/definitions/webAuthnAssertion'
...
  publicHttpError:
    type: object
    required:
    - status
    - type
    - title
    properties:
      detail:
        description: More detailed, human-readable, optional explanation of the error
        type: string
        example: User is lacking permission to access this resource
      status:
        description: HTTP status code returned for the error
        type: integer
        format: int64
        maximum: 599
        minimum: 100
        x-go-name: Code
        example: 403
      title:
        description: Short, human-readable description of the error
        type: string
        example: Forbidden
      type:
        $ref: '#/definitions/publicHttpErrorType'
  publicHttpErrorType:
    description: Type of error returned, should be used for client-side error handling
    type: string
    enum:
    - generic
    - PUSH_TOKEN_ALREADY_EXISTS
    - OLD_PUSH_TOKEN_NOT_FOUND
    - ZERO_FILE_SIZE
    - USER_DEACTIVATED
    - INVALID_PASSWORD
    - NOT_LOCAL_USER
    - TOKEN_NOT_FOUND
    - TOKEN_EXPIRED
    - USER_ALREADY_EXISTS
    - MALFORMED_TOKEN
    - LAST_AUTHENTICATED_AT_EXCEEDED
    - MISSING_SCOPES
    - validation
...
  signTransactionResponse:
    type: object
    required:
    - session_id
    - status
    properties:
      estimated_time:
        description: 预计完成时间
        type: string
        example: 2s
      session_id:
        description: 签名会话 ID
        type: string
        format: uuid
      session_token:
        description: 用于连接 Signer 的鉴权 Token
        type: string
      signer_endpoints:
        description: Signer 节点 gRPC 地址列表
        type: array
        items:
          type: string
        example:
        - 192.168.1.100:9091
      status:
        type: string
        enum:
        - pending
        - signing
        - completed
        - failed
        example: pending
  transaction:
    type: object
    properties:
      block_number:
        type: integer
      from:
        type: string
        example: 0x...
      status:
        type: string
        enum:
        - pending
        - confirmed
        - failed
        example: confirmed
      timestamp:
        type: string
        format: date-time
      to:
        type: string
        example: 0x...
      tx_hash:
        type: string
        example: 0x...
      value:
        type: string
        example: "1.5"
  transactionsResponse:
    type: object
    required:
    - transactions
    - total
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      transactions:
        type: array
        items:
          $ref: '#/definitions/transaction'
  walletBalanceResponse:
    type: object
    required:
    - balance
    - symbol
    - decimals
    properties:
      balance:
        description: 余额（字符串，避免精度损失）
        type: string
        example: "1.5"
      chain_type:
        type: string
        example: ethereum
      decimals:
        type: integer
        example: 18
      symbol:
        type: string
        example: ETH
  walletResponse:
    type: object
    required:
    - wallet_id
    - address
    - public_key
    - chain_type
    properties:
      address:
        type: string
        example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
      algorithm:
        type: string
        example: ECDSA
      chain_type:
        type: string
        example: ethereum
      created_at:
        type: string
        format: date-time
      curve:
        type: string
        example: secp256k1
      public_key:
        description: 公钥（hex）
        type: string
        example: 0x04...
      threshold:
        type: integer
        example: 2
      total_nodes:
        type: integer
        example: 2
      wallet_id:
        type: string
        format: uuid
  walletSummary:
    type: object
    properties:
      address:
        type: string
        example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
      algorithm:
        type: string
        example: ECDSA
      chain_type:
        type: string
        example: ethereum
      created_at:
        type: string
        format: date-time
      curve:
        type: string
        example: secp256k1
      public_key:
        type: string
        example: 0x04...
      wallet_id:
        type: string
        format: uuid
  webAuthnAssertion:
    type: object
    required:
    - credential_id
    - authenticator_data
    - client_data_json
    - signature
    properties:
      authenticator_data:
        description: Authenticator Data (base64)
        type: string
        format: byte
      client_data_json:
        description: Client Data JSON (base64)
        type: string
        format: byte
      credential_id:
        description: WebAuthn Credential ID (base64)
        type: string
        format: byte
        example: Y3JlZGVudGlhbC1pZA==
      signature:
        description: Signature (base64)
        type: string
        format: byte
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query