# internal/types/create_wallet_response.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// CreateWalletResponse create wallet response
//
// swagger:model createWalletResponse
type CreateWalletResponse struct {

	// DKG 会话 ID
	// Required: true
	// Format: uuid
	DkgSessionID *strfmt.UUID `json:"dkg_session_id"`

	// 用于连接 Signer 的鉴权 Token
	SessionToken string `json:"session_token,omitempty"`

	// Signer 节点 gRPC 地址列表
	// Example: ["192.168.1.100:9091"]
	SignerEndpoints []string `json:"signer_endpoints"`

	// 会话状态
	// Example: pending
	// Required: true
	// Enum: [pending running completed failed]
	Status *string `json:"status"`

	// 钱包 ID
	// Required: true
	// Format: uuid
	WalletID *strfmt.UUID `json:"wallet_id"`
}

// Validate validates this create wallet response
func (m *CreateWalletResponse) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateDkgSessionID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateStatus(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateWalletID(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *CreateWalletResponse) validateDkgSessionID(formats strfmt.Registry) error {

	if err := validate.Required("dkg_session_id", "body", m.DkgSessionID); err != nil {
		return err
	}

	if err := validate.FormatOf("dkg_session_id", "body", "uuid", m.DkgSessionID.String(), formats); err != nil {
		return err
	}

	return nil
}

var createWalletResponseTypeStatusPropEnum []interface{}

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["pending","running","completed","failed"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		createWalletResponseTypeStatusPropEnum = append(createWalletResponseTypeStatusPropEnum, v)
	}
}

const (

	// CreateWalletResponseStatusPending captures enum value "pending"
	CreateWalletResponseStatusPending string = "pending"

	// CreateWalletResponseStatusRunning captures enum value "running"
	CreateWalletResponseStatusRunning string = "running"

	// CreateWalletResponseStatusCompleted captures enum value "completed"
	CreateWalletResponseStatusCompleted string = "completed"

	// CreateWalletResponseStatusFailed captures enum value "failed"
	CreateWalletResponseStatusFailed string = "failed"
)

// prop value enum
func (m *CreateWalletResponse) validateStatusEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, createWalletResponseTypeStatusPropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *CreateWalletResponse) validateStatus(formats strfmt.Registry) error {

	if err := validate.Required("status", "body", m.Status); err != nil {
		return err
	}

	// value enum
	if err := m.validateStatusEnum("status", "body", *m.Status); err != nil {
		return err
	}

	return nil
}

func (m *CreateWalletResponse) validateWalletID(formats strfmt.Registry) error {

	if err := validate.Required("wallet_id", "body", m.WalletID); err != nil {
		return err
	}

	if err := validate.FormatOf("wallet_id", "body", "uuid", m.WalletID.String(), formats); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this create wallet response based on context it is used
func (m *CreateWalletResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateWalletResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateWalletResponse) UnmarshalBinary(b []byte) error {
	var res CreateWalletResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}